split consts
