+++ src/packages/user/user/repository/detail/index.ts
@@ -1,5 +1,5 @@
 export { UmbUserDetailRepository } from './user-detail.repository.js';
-export { UMB_USER_DETAIL_STORE_CONTEXT } from './user-detail.store.js';
+export { UMB_USER_DETAIL_STORE_CONTEXT } from './user-detail.store.token.js';
 export * from './constants.js';
 
 export type { UmbUserDetailStore } from './user-detail.store.js';
