+++ src/packages/user/user/repository/detail/user-detail.repository.ts
@@ -1,7 +1,7 @@
 import type { UmbUserDetailModel } from '../../types.js';
 import type { UmbUserDetailDataSource } from './types.js';
 import { UmbUserServerDataSource } from './user-detail.server.data-source.js';
-import { UMB_USER_DETAIL_STORE_CONTEXT } from './user-detail.store.js';
+import { UMB_USER_DETAIL_STORE_CONTEXT } from './user-detail.store.token.js';
 import type { UmbControllerHost } from '@umbraco-cms/backoffice/controller-api';
 import { UmbDetailRepositoryBase } from '@umbraco-cms/backoffice/repository';
