+++ src/packages/user/user/repository/disable/constants.ts
@@ -1,0 +1,1 @@
+export const UMB_DISABLE_USER_REPOSITORY_ALIAS = 'Umb.Repository.User.Disable';
