+++ src/packages/user/user/repository/disable/index.ts
@@ -1,2 +1,2 @@
 export { UmbDisableUserRepository } from './disable-user.repository.js';
-export { UMB_DISABLE_USER_REPOSITORY_ALIAS } from './manifests.js';
+export { UMB_DISABLE_USER_REPOSITORY_ALIAS } from './constants.js';
