+++ src/packages/user/user/repository/disable/manifests.ts
@@ -1,6 +1,6 @@
+import { UMB_DISABLE_USER_REPOSITORY_ALIAS } from './constants.js';
 import type { ManifestRepository, ManifestTypes } from '@umbraco-cms/backoffice/extension-registry';
 
-export const UMB_DISABLE_USER_REPOSITORY_ALIAS = 'Umb.Repository.User.Disable';
 const disableRepository: ManifestRepository = {
 	type: 'repository',
 	alias: UMB_DISABLE_USER_REPOSITORY_ALIAS,
