+++ src/packages/user/user/repository/enable/constants.ts
@@ -1,0 +1,1 @@
+export const UMB_ENABLE_USER_REPOSITORY_ALIAS = 'Umb.Repository.User.Enable';
