+++ src/packages/user/user/repository/enable/index.ts
@@ -1,2 +1,2 @@
 export { UmbEnableUserRepository } from './enable-user.repository.js';
-export { UMB_ENABLE_USER_REPOSITORY_ALIAS } from './manifests.js';
+export { UMB_ENABLE_USER_REPOSITORY_ALIAS } from './constants.js';
