+++ src/packages/user/user/repository/item/index.ts
@@ -1,5 +1,5 @@
 export { UmbUserItemRepository } from './user-item.repository.js';
-export { UMB_USER_ITEM_STORE_CONTEXT } from './user-item.store.js';
-export { UMB_USER_ITEM_REPOSITORY_ALIAS, UMB_USER_ITEM_STORE_ALIAS } from './manifests.js';
+export { UMB_USER_ITEM_STORE_CONTEXT } from './user-item.store.token.js';
+export * from './constants.js';
 
 export type { UmbUserItemModel } from './types.js';
