+++ src/packages/user/user/repository/item/manifests.ts
@@ -1,7 +1,5 @@
+import { UMB_USER_ITEM_REPOSITORY_ALIAS, UMB_USER_ITEM_STORE_ALIAS } from './constants.js';
 import type { ManifestItemStore, ManifestRepository, ManifestTypes } from '@umbraco-cms/backoffice/extension-registry';
-
-export const UMB_USER_ITEM_REPOSITORY_ALIAS = 'Umb.Repository.User.Item';
-export const UMB_USER_ITEM_STORE_ALIAS = 'Umb.ItemStore.User';
 
 const itemRepository: ManifestRepository = {
 	type: 'repository',
