+++ src/packages/user/user/repository/item/user-item.store.ts
@@ -1,5 +1,5 @@
 import type { UmbUserItemModel } from './types.js';
-import { UmbContextToken } from '@umbraco-cms/backoffice/context-api';
+import { UMB_USER_ITEM_STORE_CONTEXT } from './user-item.store.token.js';
 import type { UmbControllerHost } from '@umbraco-cms/backoffice/controller-api';
 import { UmbItemStoreBase } from '@umbraco-cms/backoffice/store';
 
@@ -23,5 +23,3 @@ export class UmbUserItemStore extends UmbItemStoreBase<UmbUserItemModel> {
 }
 
 export default UmbUserItemStore;
-
-export const UMB_USER_ITEM_STORE_CONTEXT = new UmbContextToken<UmbUserItemStore>('UmbUserItemStore');
